+++ example/demo/demo.dart
@@ -74,6 +74,7 @@ class _HomePageState extends State<HomePage> {
         pinStatus: _leftPinned ? PinStatus.left : PinStatus.none,
         leading: const Icon(Icons.person, size: 16),
         name: 'Name',
+        cellFocusTraversalEnabled: true,
         width: 100,
         //cellValue: (params) => params.data.name,
         cellWidget: (param) => TextField(),